EduCareer-Navigator/project/advisor_agent/requirements.txt
# Advisor Agent Requirements
# AI and Language Processing
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
openai>=1.0.0

# Data Validation and Type Hints
pydantic>=2.0.0
typing-extensions>=4.5.0

# Web Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Text Processing and Regular Expressions
regex>=2023.0.0

# Database (if using local storage)
sqlite3  # Built-in with Python

# Utilities
python-dotenv>=1.0.0
pathlib2>=2.3.0  # For Python < 3.4 compatibility
uuid  # Built-in with Python

# Development and Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Jupyter (for development)
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0 